Remove incorrectly applied or doubled ASF license headers
The source header policy[1] says that you should not prepend the Apache
License when distributing work that is derived from projects under other
licenses. We have bits and pieces of code from LevelDB and RocksDB (BSD
license) which we'd previously prepended the ASF license to, but instead
should just keep the original license.

I briefly checked the 'git blame' on these files to see the extent to which
we'd modified them, and in some cases it was around 40-50% of the lines of
code, but none was rewritten to the extent that it's worth starting a discussion
on the subtleties of "major" vs "minor" modifications.

This also removes some cases in which we'd applied the Apache license twice
to the same piece of code -- in all cases in code we'd imported from Impala
at some point along the line after it was Apache licensed.

Change-Id: I8d8e781b672764a84792203b09e85c67e535dd11
Reviewed-on: http://gerrit.cloudera.org:8080/2484
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins