Fix flakiness in remote_bootstrap-itest
This test deleted a tablet, and then checked to see that the tablet data state
was TABLET_DATA_COPYING. However, it was possible for the remote bootstrap to
complete so quickly that it would already be TABLET_DATA_READY by the time
we checked, making the test flaky.

This changes the check to allow either COPYING or READY states.

Without the patch, it was around 8% flaky[1]. With the patch, it's not flaky[2].

[1] http://dist-test.cloudera.org/job?job_id=todd.1457557419.27817
[2] http://dist-test.cloudera.org/job?job_id=todd.1457559502.30323

Change-Id: I07b53a4209316b6a543f4d3d39ff17bed3890dd9
Reviewed-on: http://gerrit.cloudera.org:8080/2505
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>