client-test: try to de-flake TestLastErrorEmbeddedInScanTimeoutStatus
This test seems to be flaky on the test dashboard, though I couldn't
reproduce it on dist-test (perhaps due to different slave hardware).

Regardless, it seems like there's a race -- we inject 1 second of
latency into the tablet server startup, and then set a 1-second
timeout on our scan. So, maybe our scan can sometimes succeed within
the target time in some cases.

This bumps the latency to 5 seconds to try to make the intended timeout
more of a sure thing.

Change-Id: I63c3e31a737b11ad3c9104d2f77111d17e9a51b7
Reviewed-on: http://gerrit.cloudera.org:8080/2506
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins