KUDU-418 (part 1): enforce that we don't support changing TS hosts/ports
This adds a check in the master that attempts to reject tablet servers which
have changed host/port. We don't currently support this, so allowing them to
register currently just ends up hitting other more hard-to-diagnose issues
later on.

This check will end up being rolled back when we actually implement a proper
fix for KUDU-418. In the meantime, this would have helped me troubleshoot some
issues with client-test where we hit this issue.

Change-Id: I16d9a91059b7dde63a2cac71e1fde0228b6a9f3d
Reviewed-on: http://gerrit.cloudera.org:8080/2513
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins