Improve site build, add a blog
Since we plan to move off of github pages towards the ASF site publishing, we
now need to start worrying about running Jekyll ourselves. This adds a new
script to this repository that acts as a wrapper which installs bundler uses
that to install a locked version of Jekyll. This slightly duplicates some
scripting which is in the main Kudu repository, but that scripting is more
focused around building documentation (eg installing the asciidoctor gem)
whereas this one is focused on the site build.

This also updates Jekyll to a newer version which has better support for
blogging. This still isn't the newest version, because the newest requires ruby
2.0, which apparently is not installed by default on a lot of systems. If we
end up needing some feature that is missing from Jekyll 2.5, we can update at
that ponit.

In terms of actual site content, this adds a basic blog, and two simple
back-dated posts for our two most recent releases. I used these as initial
filler content so that we don't have an empty blog.

In order to make room on the top nav for the blog, I merged the 'Contributing'
section into the 'Community' section. Longer term I think we probably need to
add some dropdown menus from the top nav or something, but I didn't want to
figure out the HTML for that quite yet. I think putting 'contributing' under
'community' makes reasonable sense.

Change-Id: Ib9afca4e9b22923b96fd4ec9f5bfafd3e6ddd796
Reviewed-on: http://gerrit.cloudera.org:8080/2542
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>