Fix Spark shutdown
The Kudu client was creating Netty timers with non-daemon threads, which
prevented the Spark shell JVM from beginning the shutdown process. This commit
also adds a shutdown hook to the Spark connector so that when the shell process
exits it will properly shut down the Kudu client which prevents batched but
unflushed writes from being lost.

See
http://apache-spark-developers-list.1001551.n3.nabble.com/graceful-shutdown-in-external-data-sources-td16684.html
for more discussion.

Change-Id: I30a64ec5eb30d70361204646523c9947d88c251f
Reviewed-on: http://gerrit.cloudera.org:8080/2571
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins