codegen: set JIT target to running host
This slight tweak to codegen tells the JIT engine to generate code for the
currently running host.

I looked at the generated assembly from codegen-test, and it's unchanged for
integer projections. When I added a string column to the projection, this patch
saved two instructions in the generated code by using some AVX instructions.

Change-Id: I49c26696db0bfd819f918aec27ff5d65a55dcc93
Reviewed-on: http://gerrit.cloudera.org:8080/2653
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>