KUDU-1396. Fix index building when an individual value overflows block
This fixes an infinite recursion when building the index for a cfile
in the case that a single value is larger than the configured index
block size. Previously, we would see that the block was full, propagate
the key upwards, see that the new block was full, and continue until
we crashed.

The fix is relatively simple: we need to make sure that an index block
has at least two elements before we propagate upwards.

This adds a new cfile test which writes strings that are larger than
64KB.

Change-Id: Ic27485bc0b2151e7c182c36e7ad6258dc549517b
Reviewed-on: http://gerrit.cloudera.org:8080/2725
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins