[c++-client] abstract scanner configuration
Unlike the Java client, the c++ client configures scan options directly on the
KuduScanner. This commit adds an internal class, ScanConfiguration, that takes
over responsibility for configuring scans from the KuduScanner.  KuduScanner is
updated to hold a ScanConfiguration internally. The ScanConfiguration will be
used internally in the forthcoming scan token builder API.

Change-Id: Id3d5b27f4a0ae6cd1500f17f96b7043590affd92
Reviewed-on: http://gerrit.cloudera.org:8080/2726
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>