Fix license headers for util/memory/overwrite.{cc,h}
One had a mistaken Cloudera copyright. The other had a mistaken Google
copyright. I traced back the git log and found that I added the current
implementation of the function in this file in
f5546effc66281f48f8c2625a30d3ab9fb50fa75, so it's fine to license to
ASF rather than Google.

Thanks to Justin Mclean for the spot.

Change-Id: I3387a0842356ceef9b76a5242206d6f5021bbb58
Reviewed-on: http://gerrit.cloudera.org:8080/2729
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins