KUDU-1409. Make krpc call timeouts more resistant to process or reactor pauses
In stress testing Impala on Kudu I've seen various RPC timeouts that turn out
to be due to pauses on the client side. In particular, scenarios like
IMPALA-2800[1] can cause the memory allocator inside Impala to block for
several seconds, and that might cause us to think we missed a timeout.

This adds a little workaround in our libev timeout handling code. The
full description of the workaround can be found in connection.cc.

[1] https://issues.cloudera.org/browse/IMPALA-2800

Change-Id: I7bff0bc1573a059f12be8bd3f46e301275e78392
Reviewed-on: http://gerrit.cloudera.org:8080/2745
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>