KUDU-1410 (part 2). rpc: clean up generated RPC service code
This changes the generated RPC service code to extract the actual
handling logic into a non-generated base class. The generated
code now just generates a map of method name to MethodInfo, each
of which contains the requisite information to handle an RPC.

This results in a small reduction in lines of non-generated code,
and a larger reduction in lines of generated code.

Furthermore, this refactor makes it easier to make improvements and
further cleanups on the RPC server side, since most changes now
won't need to affect the code generator.

Change-Id: Id7b42a5ba1de59315660a79b125e70938c3a1b2e
Reviewed-on: http://gerrit.cloudera.org:8080/2795
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins