KUDU-1307 [master] support tables with range partition bounds
This commit adds support to the Kudu master for creating tables with range
partition bounds. Bounds are used to create gaps in the range partition that are
not covered by a tablet. A new master feature flag has been added so that when
client support is added for creating tables with partition bounds, old masters
will not silently ignore the bounds. Adding support to the clients is left to a
follow up commit.

Change-Id: I04a7ed3e85c993fdbcb313a419f15031103736ce
Reviewed-on: http://gerrit.cloudera.org:8080/2806
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>