[java client] Add a way to restart processes
This patch adds more tracking of the processes started by MiniKuduCluster.
Mainly we needed to record the command lines used to start the masters
and tablet servers.

This patch adds a new unit test for the mini cluster, before we only relied
on the fact that everything was using it so therefore it was tested.

Finally getUniqueLocalhost was moved to TestUtils so that findFreePort
could also use it. The reason is that the processes on OSX were binding
to 0.0.0.0 but findFreePort was using a different hostnames in
ServerSocket and so the behavior was broken. This was noticed when a TS
leaked after a test and new tests were failing because the TS was already
bound to a port we thought to be free.

Change-Id: I065917cd219e1dd8fa93fed9ca36d58a3372935d
Reviewed-on: http://gerrit.cloudera.org:8080/2828
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Dan Burkert <dan@cloudera.com>