Add a 'site_tool publish' command to publish to the ASF site
As part of moving to ASF infrastructure, we can't rely on github pages in the
future. This means we need to actually publish the site to static HTML which
gets committed into the ASF's gitpubsub repository.

This is of course slightly error prone, so it's important to have a nice script
to make it easy for anyone to publish the latest version of the site.
This patch introduces such a command to the 'site_tool' script.

The main idea here is that we check out the target repository into '_publish',
build the site, and generate a reasonable commit message which explains the
changes since the prior commit.

The current version doesn't go so far as to automate the push itself, but
instead just gives instructions to the user on how to do so. We could easily
automate the whole thing at a later point in time, but it was easier to
test in a non-destructive form.

Change-Id: Icc4f21bbad06cf38322d92a7fe3e31dc5a8b9f9a
Reviewed-on: http://gerrit.cloudera.org:8080/2856
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Mike Percy <mpercy@apache.org>