rpc: avoid crashing if the server generates an invalid RPC response
This prevents a server-wide crash in the case that an RPC handler tries to
respond with an invalid protobuf (e.g. missing a required field).

We used to have a DFATAL in this case, which would log an error, but then
proceed to crash immediately thereafter because the response buffer was empty.

The new behavior is for the server side to log the error, and then send
the partial protobuf back to the client anyway. The client will see
the same issue upon parsing the response, and generate an error to
the caller indicating the missing fields. This has the benefit that
the error will be logged on both sides, and that the server won't crash.

This also changes the behavior on the client side when trying to send
an invalid RPC. Again, we send it even though it's incomplete, but
expect the server to reject it.

Change-Id: Ia7845388b1f520158e554548d668836d40b10ca6
Reviewed-on: http://gerrit.cloudera.org:8080/2914
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>