KUDU-1433: check stat validity before dumping mm ops
Some mm ops don't cache their stats, and in doing so, can yield invalid
stats following a call to FindBestOp(). Users of these stats need to be
aware that they may be invalid, and to protect themselves accordingly. In
this particular case, I think it makes more sense to provide zeroed stats
than to skip the op altogether.

After modifying the PeriodicWebUIChecker to access /maintenance-manager,
linked_list-test failed 100% of the time. With the fix, it doesn't fail.

Change-Id: I4c363c8b1adb9724a2c1deb09a27a7d89d84c270
Reviewed-on: http://gerrit.cloudera.org:8080/2922
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>