Add a request tracker to track client rpc sequence numbers
This adds a new component to the rpc subsystem: the RequestTracker. This is inspired
by the RequestTracker in "Implementing Linearizability at Large Scale and Low Latency"
by Colin Lee et al. and is responsible for generating new sequence numbers for rpcs
and to keep track of the incomplete ones so that we're able to implement exactly-once
semantics for certain rpcs.

Change-Id: I23201625ca02f244dc94205d88dabc01608de471
Reviewed-on: http://gerrit.cloudera.org:8080/3078
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>