[java client] Add a method to return rows in order
This patch adds an unstable package private API on the scanners to request
the rows to be returned sorted by primary key. This is meant to be used
by more advanced users that understand what they are doing. It will remain
package private until we start internally checking that if the table is
hash partitioned, all hash columns have equality predicates.

The new test changes how ITScannerMultiTablet behaves. It now inserts more
rows and does so randomly, into a hash-partitioned table. That, plus a new
config for TS to flush after 1MB, makes it possible to test this new method.

FWIW the change to 1MB flush applies to all the Java client tests. It could
be made configurable but I haven't found a need for it yet.

Change-Id: I865e28bddd945111ee159b6f2715a8629b75743b
Reviewed-on: http://gerrit.cloudera.org:8080/2951
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins