Refactor cache interface to handle allocation of values and keys together
This changes the cache interface so that the cache itself now
allocates the LRUHandle, the key, and the value all together.
The insertion process is now "two-phase": first, we allocate a new
cache entry with the appropriate size, then fill in the value,
and then we pass the entry back to the cache for insertion.

Because the cache now owns all of the memory, the "deleter"
callback is no longer necessary to free cache values. Instead,
it's just an "eviction" callback, since we still need to
take action on eviction for the code cache (and for tests).

The point of this refactor is to enable the persistent NVM cache,
which needs to allocate LRUHandles on the heap, but the keys
and values in persistent memory.

Change-Id: Iedeb24c0233bf062a0a6450a9fd3a7c57499144f
Reviewed-on: http://gerrit.cloudera.org:8080/2957
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>