Bump guava version to 18.0
We use the javax thread safety annotations from the jsr305 jar which the
previous Guava version had a compile scope dependency on. Guava has since made
that dependency optional since it only contains annotations, so we have to add
an optional dependency on it as well.

The new guava version conflicts with the version used by MR, so the
kudu-mapreduce unit tests were converted into integration tests, so that the
tests run after the shading step.

Change-Id: I6d18c96d94544b493d7ca92eff6205d58f11cfdc
Reviewed-on: http://gerrit.cloudera.org:8080/2972
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>