Short-circuit CheckRowDeleted if there are no deletes in DeltaFile
For every mutation, we need to check whether the existing row has been deleted.
In the case that there are delta files, this causes a deltafile seek.

By looking at the deltafile stats up front, we can tell whether there are any
deletes at all, and if there aren't, we know that the row in question hasn't
been deleted.

Looking at a profile of a heavy update workload on a large table,
CheckRowDeleted() and its descendents were taking about 25% of the total CPU,
so this should be a substantial win.

Change-Id: Iaa35bab3f9a9fc82149d11821733e87a7aaf5e7a
Reviewed-on: http://gerrit.cloudera.org:8080/2975
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins