Add an all-virtual ServerPicker class and a meta cache backed implementation
This pulls out the replica picking code from WriteRpc and into an implementation
of the new ServerPicker, all-virtual class.

The new server picker implementation, which is backed by the meta cache, fully
readies the replica before handing it over to WriteRpc. This makes the writing
logic simpler (less callbacks) and makes the replica picking code possibly reusable.

Again, this is just a refactor of existing code so no new tests were added.

Change-Id: Id43db316606cb807ec0019c79b3bdf76fa509fe5
Reviewed-on: http://gerrit.cloudera.org:8080/3017
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>