Add a cached boolean whether DMS is empty
On servers which store a couple TB of data, I see a good portion of a core
being burned in DeltaTracker::DeltaMemStoresEmpty(). Even the component lock
is a rw-lock, the atomic instructions necessary to acquire the lock cost a lot.

This adds a boolean member to DeltaTracker which is set on update to
try to reduce this CPU load.

Change-Id: I721ee74a112d968a33a1a82419b3268c08b67c79
Reviewed-on: http://gerrit.cloudera.org:8080/3004
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>