Add some more latency-related trace counters
In looking at various RPCs in the new /rpcz profiling mechanism, I noted
that a few sources of latency were not well accounted for. This patch adds
latency counters for the time spent starting new threads (eg if a threadpool
spawns a new worker) and for time spent fsyncing data on disk.

A handy new utility class is used to make this kind of counter easy to add.

Change-Id: Ie601e914c22b4ca86dc25f4be0c6ec5ef0e94404
Reviewed-on: http://gerrit.cloudera.org:8080/3008
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins