KUDU-1444. Get resource metrics of a scan.
This patch supports to get the resource metrics of a scan in client side. The
resource metrics will be sent back to client in every scan RPC response. This
is useful for impala to show these stats in a query profile.

For now, the resource metrics only contains cfile_cache_miss_bytes and
cfile_cache_hit_bytes. We may add more in the future as needed.

Change-Id: Iedaf570a7601651c93275ae0a8565f1e33da842d
Reviewed-on: http://gerrit.cloudera.org:8080/3013
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>