Make the meta cache return Status::NetworkError if no addresses were found
This adds a new flag 'fail_dns_resolution' to net-util.cc so that we can fail
dns resolution on purpose and makes it so that the meta cache returns Status:NetworkError
in this case, instead of NotFound as before.

This also adds a test that makes sure that we get Status::NetworkError() back
if we set the flag and that normal resolution resumes once it's set to the default
value.

Changing this error type allows to have a clearer error handling path as there is
one less status type to handle.

Change-Id: I95b553f1db1b76f9e8117b6628c7bf64013d1ad9
Reviewed-on: http://gerrit.cloudera.org:8080/3027
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>