log: refactor log reading into an iterator-like interface
This adds a new interface to iterate over a WAL one entry at a time. This
interface is both more natural to use and more performant: by not materializing
lots of entries in memory at the same time, the allocator thread cache should
be kept much hotter and a smaller amount of memory can be reused, keeping CPU
caches hotter.

Currently, this is just a refactor of the internals, with the existing
ReadEntries() method changed to use the iterator interface. Follow-on
patches will change call sites to use the iterator instead.

Change-Id: I5727951664600b7a591a6c892ea1e62469ceb109
Reviewed-on: http://gerrit.cloudera.org:8080/3038
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>